requirements.txt
snowflake-connector-python==2.7.9
requests~=2.28.1
PyYAML~=6.0